Add Overcooked rendering functionality
- Introduced overcooked_render.h to encapsulate all rendering and texture management functions for the Overcooked environment.
- Implemented functions for agent color determination based on held items and texture unloading.
- Integrated rendering logic into the Overcooked environment, including window initialization, texture loading, and drawing operations for agents, items, and cooking states.
- Updated overcooked.h to include the new overcooked_render.h header, enhancing code organization and maintainability.
